% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{add_annotations}
\alias{add_annotations}
\title{Adds sample annotation information to loaded MaxQuant or IceR data}
\usage{
add_annotations(data_list, Annotations)
}
\arguments{
\item{data_list}{List object containing loaded MaxQuant or IceR data}

\item{Annotations}{Table with at least one column containing annotation information. Rows have to correspond to samples in loaded MaxQuant or IceR data. Requires same order of sampels (rows) as e.g. sampels (columns) in data_list$Protein_level$Quant_data}
}
\value{
List object containing loaded MaxQuant or IceR data extended by annotation information
}
\description{
Adds sample annotation information to loaded MaxQuant or IceR data
}
\details{
Add sample annotation to loaded MaxQuant or IceR data.
}
