% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{add_isotope_features}
\alias{add_isotope_features}
\title{Extend IceR features by expected +1-isotopic features}
\usage{
add_isotope_features(
  path_to_features,
  feature_table_file_name = "Features_aligned_merged_IceR_analysis.txt",
  min_observations = 0
)
}
\arguments{
\item{path_to_features}{Path to folder where results of align_features() are stored.}

\item{feature_table_file_name}{File name which contains align_features() results. By default is set to Features_aligned_merged_IceR_analysis.txt.}

\item{min_observations}{Specifying how many MaxQuant features had to be previously detected for monoisotopic IceR features to be extended by +1-isotopic features. By default set to 0 indicating that all for all monoisotopic IceR features a +1-isotopic feature is added.}
}
\value{
Extended list ist stored in the specified feature_table_file_name.
}
\description{
Extend IceR features by expected +1-isotopic features
}
\details{
Optional step of the IceR workflow appending the list of IceR features by respective expected +1-isotopic features. Isotopic features are expected to show an m/z shift by +1.002 Da/z relative to the monoisotopic IceR feature. Only IceR features with at least min_observations and with mean Andromeda score > 25 % quantile of all mean Andromeda scores are considered.
}
