% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{add_missed_peptides}
\alias{add_missed_peptides}
\title{Extend IceR features by expected monoisotopic features which correspond to totally missed tryptic peptides of identified proteins.}
\usage{
add_missed_peptides(
  path_to_features,
  feature_table_file_name = "Features_aligned_merged.txt",
  path_to_fasta,
  min_AA_length = 7,
  max_AA_length = 25,
  max_add_per_protein = 10,
  min_observations = 5
)
}
\arguments{
\item{path_to_features}{Path to folder where results of align_features() are stored.}

\item{feature_table_file_name}{File name which contains align_features() results. By default is set to Features_aligned_merged_IceR_analysis.txt.}

\item{path_to_fasta}{Path to fasta file used during MaxQuant data preprocessing.}

\item{min_AA_length}{Minimal length of a theoretical peptide which is considered to be added. By default set to 7.}

\item{max_AA_length}{Maximal length of a theoretical peptide which is considered to be added. By default set to 25.}

\item{max_add_per_protein}{Maximal number of potentially missed peptide features which should be added. By default set to 10.}

\item{min_observations}{Minimal number of MaxQuant features which have to fall into the expected RT- and m/z-windows of potentially missed peptide features.}
}
\value{
Extended list ist stored in the specified feature_table_file_name.
}
\description{
Extend IceR features by expected monoisotopic features which correspond to totally missed tryptic peptides of identified proteins.
}
\details{
Optional step of the IceR workflow appending the list of IceR features by potentially missed tryptic peptide IceR features. Requires installation of the Bioconductor package "cleaver". Expected m/z windows are determined based on masses of respective tryptic peptides. Expected RT windows are estimated based on random forest modelling with KyteDoolittle hydrophobicity, peptide length, peptide charges and amino acid composition as predictors and observed RTs as response factors. Potentially missed peptide (PMP) features are expected to show +2 or +3 charge.
}
