% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{align_features}
\alias{align_features}
\title{Perform alignment of pre-determined MS1-features by MaxQuant over proteomics samples either analyzed on an Orbitrap machine (e.g. Q-Exactive or Orbitrap Fusion) or a TIMS-ToF Pro}
\usage{
align_features(
  path_to_MaxQ_output,
  path_to_output,
  align_unknown = F,
  output_file_names_add = "IceR_analysis",
  mz_window = NA,
  min_mz_window = 0.001,
  RT_window = NA,
  min_RT_window = 1,
  min_num_ions_collapse = 10,
  feature_mass_deviation_collapse = 0.002,
  only_unmodified_peptides = F,
  sample_list = NA,
  remove_contaminants = T,
  MassSpec_mode = c("Orbitrap", "TIMSToF"),
  IM_window = NA,
  min_IM_window = 0.002,
  multiplicity = c(1, 2, 3),
  SILAC_settings = list(light = c(""), medium = c("Lys4", "Arg6"), heavy = c("Lys8",
    "Arg10"))
)
}
\arguments{
\item{path_to_MaxQ_output}{Path to folder containing MaxQuant outputs (txt folder containing at least allpeptides.txt, evidence.txt, peptides.txt and proteinGroups.txt)}

\item{path_to_output}{Path to folder where IceR results should be stored}

\item{align_unknown}{Boolean value indicating if only peptide features or also unsequenced features should be aligend over samples. By default set to F.}

\item{output_file_names_add}{IceR result name tag. By default IceR_analysis}

\item{mz_window}{Numeric value indicating maximal m/z deviation around a center. By default set to NA which indicates that the function determines automatically this parameter based on sd of m/z of identified peptides between samples.}

\item{min_mz_window}{Numberic value indicating how large the automatically determined m/z-window should be. By default set to 0.001 Da. Only required if m/z alignment window should be automatically determined. If set to NA, no minimal window size will be required.}

\item{RT_window}{Numeric value indicating maximal RT deviation around a center. By default set to NA which indicates that the function determines automatically this parameter based on sd of RT of identified peptides between samples.}

\item{min_RT_window}{Numberic value indicating how large the automatically determined RT-window should be. By default set to 1 min. Only required if RT alignment window should be automatically determined. If set to NA, no minimal window size will be required.}

\item{min_num_ions_collapse}{Numeric value indicating how many unsequenced MaxQuant features have to be at least detected over all samples to result in an IceR feature. Only required if align_unknown is set to T. By default set to 10.}

\item{feature_mass_deviation_collapse}{Numeric value indicating which minimal mass deviation is required to distinguish IceR features. By default set to 0.002 Da. IceR features with overlapping RT-windows and mass differences smaller than the specified value are merged.}

\item{only_unmodified_peptides}{Boolean value indicating if only unmodified peptide sequences are used for alignment. By default set to F.}

\item{sample_list}{Character vector (raw file names) listing which samples should be aligned. By default all samples occuring in MaxQuant outputs are aligned.}

\item{remove_contaminants}{Boolean value indicating if peptide features labeled as contaminants should be removed. By default set to T.}

\item{MassSpec_mode}{String being either "Orbitrap" or "TIMSToF" specifying by which type of Mass Spectrometer the data was generated. By default it expects Thermo Orbitrap data.}

\item{IM_window}{Numeric value indicating maximal ion mobility (inverse K0) deviation around a center. By default set to NA which indicates that the function determines automatically this parameter based on sd of ion mobility of identified peptides between samples.}

\item{min_IM_window}{Numberic value indicating how large the automatically determined ion mobility-window (inverse K0) should be. By default set to 0.002. Only required if ion mobility alignment window should be automatically determined. If set to NA, no minimal window size will be required.}

\item{multiplicity}{Numeric value between 1 - 3 indicating if multiple channels are mixed on MS1-level. Label-free corresponds to 1 and SILAC to 2 (light and heavy) or 3 (light, medium, heavy)}

\item{SILAC_settings}{List of 3 string vectors named light, medium and heavy. Each string vector should indicate which heavy isotope of Lys and/or Arg was used for respective channel. String vectors can contain Lys2, Lys4, Lys6, Lys8, Arg6 and Arg10.}
}
\value{
Outputs are stored in the sub-directory Temporary_files within specified output folder. MaxQuant allpeptides.txt and evidence.txt are converted to RData files. QC plots of estimated alignment windows as well as of random forest modesl and generalized additive models are stored in a QC_plots.pdf. Relevant QC data is stored in Feature_alignment_QC_data.RData. Aligned IceR features are stored in Features_aligned_merged.txt
}
\description{
Perform alignment of pre-determined MS1-features by MaxQuant over proteomics samples either analyzed on an Orbitrap machine (e.g. Q-Exactive or Orbitrap Fusion) or a TIMS-ToF Pro
}
\details{
Performs the first steps of the IceR workflow: 1) Alignment window determination if not specified. 2) Alignment of MaxQuant features into IceR features. 3) Transfer of sequence information between MaxQuant features aligned into IceR features. 4) Extraction, modelling and prediction of RT- and m/z-correction factor per IceR feature and sample.
}
