% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{BarplotsSBS}
\alias{BarplotsSBS}
\title{Generate side-by-side barplots}
\usage{
BarplotsSBS(
  Data,
  ErrbarData = NA,
  Name = "",
  ylab = "Y-axis",
  main = "Titel",
  col = "lightblue",
  AvgLine = T,
  Legends = NA,
  Legendtitle = "Legend",
  Legendpos = "topright",
  ylim = NA,
  logy = F,
  shownumbers = F,
  shownumbers_digits = 1,
  separation = T,
  horiz_line = NULL,
  margins = c(8, 4, 4, 4),
  inset = c(-0.1, 0)
)
}
\arguments{
\item{Data}{Numeric vector or table of samples in columns}

\item{ErrbarData}{Data for errorbars}

\item{Name}{Names}

\item{ylab}{Y-Axis label}

\item{main}{Plot main title}

\item{col}{Color}

\item{AvgLine}{Show average line?}

\item{Legends}{Legends}

\item{Legendtitle}{Titel for legends}

\item{Legendpos}{Legend position}

\item{ylim}{y-axis limits}

\item{logy}{Y-Axis in log-scale?}

\item{shownumbers}{Show numbers on top of bars}

\item{shownumbers_digits}{Number of digits for shown numbers}

\item{separation}{Indicate separation bars}

\item{horiz_line}{Show horizontal lines}

\item{margins}{Margins}

\item{inset}{Inset for legend}
}
\value{
Plot.
}
\description{
Generate side-by-side barplots
}
\details{
Generate side-by-side barplots
}
