% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Barplotsstacked}
\alias{Barplotsstacked}
\title{Generate stacked barplots}
\usage{
Barplotsstacked(
  Data,
  Name = "",
  ylab = "Y-axis",
  logy = F,
  main = "Titel",
  col = "lightblue",
  AvgLine = T,
  Legends = NA,
  Legendtitle = "Legend",
  Legendpos = "topright",
  shownumbers = T,
  shownumbers_total = T,
  order_groups = F,
  group_names = "",
  ylim = NULL,
  margins = c(12, 4, 4, 9),
  inset = c(-0.3, 0)
)
}
\arguments{
\item{Data}{Numeric vector or table of samples in columns}

\item{Name}{Names}

\item{ylab}{Y-Axis label}

\item{logy}{Y-Axis in log-scale?}

\item{main}{Plot main title}

\item{col}{Color}

\item{AvgLine}{Show average line?}

\item{Legends}{Legends}

\item{Legendtitle}{Titel for legends}

\item{Legendpos}{Legend position}

\item{shownumbers}{Show numbers on top of bars}

\item{shownumbers_total}{Show total numbers}

\item{order_groups}{Order groups}

\item{group_names}{Group names}

\item{ylim}{y-axis limits}

\item{margins}{Margins}

\item{inset}{Inset for legend}
}
\value{
Plot.
}
\description{
Generate stacked barplots
}
\details{
Generate stacked barplots
}
