% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{casedetect-class}
\alias{casedetect-class}
\title{Case-Detection Class}
\description{
Case-Detection Class
}
\details{
Class for case-detection data where the true
prevalence of cases is not known. Instead, we just have the
number of cases that were recorded.
}
