% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{cea}
\alias{cea}
\title{Cost-effectiveness analysis}
\usage{
cea(formula, data = NULL)
}
\arguments{
\item{formula}{Left-hand side contains (1) clinical efficacy and (2) cost.
Right-hand side contains (1) intervetion identifier and an arbitrary number
of analysis identifiers.  Formula parts on either side are separated by \code{|}.}

\item{data}{Optional data.frame containing variables defined in formula.  If not
specified, R will look for the variables in the calling environment.}
}
\value{
an object of of class CEA containing the following elements
\item{call}{A copy of the function call, for use in post-processing.}
\item{formula}{The formula used to define the analyses.}
\item{data}{A data.frame containing the results.}
}
\description{
Cost-effectiveness analysis results for a given set of scenarios.
Each scenario contains clinical efficacy and costs for each comparator included.
The output is stored in an object of class "CEA".  Generic methods are provided
for post-processing the results.
}

