% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{coinf_plot}
\alias{coinf_plot}
\title{Perform a bootstrap and plot the expected versus observed infection distributions}
\usage{
coinf_plot(reps = 5000, probs, levels, total, real, plot = TRUE,
  quantiles = c(0.025, 0.975), ...)
}
\arguments{
\item{plot}{Boolean for default plotting the bootsrap results. Default = TRUE}

\item{quantiles}{Vector of length 2 for the quantiles used. Default = `c(0.025, 0.975)`}

\item{boot_iter}{Bootstrap iterations. Default = 10000}
}
\value{
ggplot object
}
\description{
Perform a bootstrap and plot the expected versus observed infection distributions
}
\details{
\enumerate{
       \item Bostrap from the fitted multinomial probabilities
       \item Plot the simulated densities
       \item Compare to the observed data
  }
}
