% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{compare_general_numbers}
\alias{compare_general_numbers}
\title{Compare general numbers between data sets}
\usage{
compare_general_numbers(
  list_of_data_lists,
  colors = c("darkgrey", "chocolate3"),
  Legendpos = "top",
  margins = c(12, 4, 4, 9),
  inset = c(-0.435, 0)
)
}
\arguments{
\item{list_of_data_lists}{List object containing lists of loaded MaxQuant or IceR data}

\item{colors}{Colors for data sets to be compared. By default two colors are specified ("darkgrey","chocolate3")}

\item{Legendpos}{Location of legend, By default set to "top"}

\item{margins}{Margins aroung plot area. By default set to c(12,4,4,9)}

\item{inset}{Inset of legend. By default set to c(-0.435,0)}
}
\value{
Comparison plots
}
\description{
Compare general numbers between data sets
}
\details{
Compare general numbers between data sets
}
