% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cooccurence_mle}
\alias{cooccurence_mle}
\title{Co-occurence MLE}
\usage{
cooccurence_mle(
  obj,
  density_func = independent,
  max_moi = 25,
  poisson = FALSE,
  size = 100,
  lower = NULL,
  upper = NULL,
  boot_iter = 10000,
  plot = TRUE,
  quantiles = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{density_func}{Function that calculates the multinomial distribution
describing the data. Default = \code{independent}}

\item{max_moi}{Maxuimum infction composition explored. Default = 25.}

\item{poisson}{Logical for determining if we use a poisson distribution to
describe the numer of infection occurences. Default = `FALSE`, which means
a negative binomial is used. `}

\item{size}{Starting value for the negative binomial size parameter.
Default = 100.}

\item{lower}{Vector of lower bounds used in fitting. Default = NULL, which will
create a vector with 0.0001 for each frequency and 0.1 for mu (and 1 for size).}

\item{upper}{Vector of upper bounds used in fitting. Default = NULL, which will
create a vector with 0.9999 for each frequency and 100 for mu (and 500 for size).}

\item{boot_iter}{Bootstrap iterations. Default = 10000}

\item{plot}{Boolean for default plotting the bootsrap results. Default = TRUE}

\item{quantiles}{Vector of length 2 for the quantiles used. Default = `c(0.025, 0.975)`}

\item{data}{Observed data. Either as a named vector or a 2 column
\code{\link{data.frame}}. See examples for more info.}
}
\value{
A list containing the estimated parameters.
}
\description{
Maximum likelihood estimation for co-infection dynamics
}
\details{
Estimates the maximum likelihood population frequency of each species
  (i.e. the names of the observed entities in our data) and the mean number
  of infections given our observed data. These estimates are used to estimate
  the multinomial probability distribution.
}
