% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{densityplots}
\alias{densityplots}
\title{Density plots}
\usage{
densityplots(
  datafcs,
  names = NA,
  col = NA,
  main = "",
  xlab = "",
  xlim = NA,
  lwd = 2
)
}
\arguments{
\item{datafcs}{Data.frame}

\item{names}{Names per samples in columns}

\item{col}{Colors per sample in columns}

\item{main}{Plot main title}

\item{xlab}{X-Axis label}

\item{xlim}{X-axis limits}

\item{lwd}{Line width}
}
\value{
Plot.
}
\description{
Density plots
}
\details{
Plot density plots for data.frame with columns = samples and rows = observations
}
