% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{determine_general_numbers}
\alias{determine_general_numbers}
\title{Determine general identification and quantification numbers}
\usage{
determine_general_numbers(data_list)
}
\arguments{
\item{data_list}{List object containing loaded MaxQuant or IceR data}
}
\value{
List object containing loaded MaxQuant or IceR data extended with general number information.
}
\description{
Determine general identification and quantification numbers
}
\details{
Determine general numbers of the data set like number of proteins, number of peptides and number of missing values
}
