% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perms.R
\name{generate_composition_levels}
\alias{generate_composition_levels}
\title{Creates composition types from a vector of probabilities}
\usage{
generate_composition_levels(probs)
}
\arguments{
\item{probs}{Named vector of probabilities}
}
\value{
Returns a named vector of 0s.
}
\description{
Creates composition types from a vector of probabilities
}
\details{
Creates a named vector of all combinations for the names
  in probs, separated by a "/" and sorted by character length
  and then alphabetically
}
\examples{
\dontrun{

probs <- c("a" = 0.5, "b" = 0.3, "c" = 0.2)
icer:::generate_composition_levels(probs = probs)

#> a     b     c   a/b   a/c   b/c a/b/c
#> 0     0     0     0     0     0     0

}
}
