% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_join_comp_moi_prob}
\alias{get_join_comp_moi_prob}
\title{Get joint probability of composition and moi}
\usage{
get_join_comp_moi_prob(obj, multi_dens, w)
}
\arguments{
\item{obj}{S4 object;}

\item{multi_dens}{parobability matrix; each row gives the probability density of
each infection type for a given multiplicity of infection.}

\item{w}{numeric vector; probability (or weight) of each MOI scenario that was
calculated by `moi_probabilities`}
}
\description{
Get joint probability of composition and moi
}
