% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Heatmap}
\alias{Heatmap}
\title{Heatmap visualization of data}
\usage{
Heatmap(
  data,
  annotation,
  annotation_name = "Anno",
  main = "Heatmap",
  colors_annotation,
  colors_plot = NA,
  color_breaks = NA,
  hclust_col = T,
  hclust_row = T,
  show_rownames = T,
  show_colnames = F,
  fontsize_rows = 10,
  interactive = F,
  row_labels = NULL
)
}
\arguments{
\item{data}{Data.frame}

\item{annotation}{Conditions per samples in columns of data}

\item{annotation_name}{Displayed name for annotation}

\item{main}{Plot main title}

\item{colors_annotation}{Annotation colors}

\item{colors_plot}{Colors for heatmap}

\item{color_breaks}{Color breaks for heatmap}

\item{hclust_col}{Cluster cols?}

\item{hclust_row}{Cluster rows?}

\item{show_rownames}{Show rownames?}

\item{show_colnames}{Show colnames?}

\item{fontsize_rows}{Fontsize of rows}

\item{interactive}{Interactive heatmap}

\item{row_labels}{Special rowlabels}
}
\value{
Plot.
}
\description{
Heatmap visualization of data
}
\details{
Heatmap visualization of data
}
