% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceR.R
\docType{package}
\name{iceR-Package}
\alias{iceR-Package}
\alias{iceR-Package-package}
\title{iceR: Post-processing CEA results}
\description{
The iceR package provides objects and methods for post-processing the
the results of cost-effectiveness analyses, including tools for reporting
results of deterministic and probabilistic sensitivity analyses.
}
\section{CEA}{

}

\section{deterministic}{

Methods for post-processing the results of deterministic sensitivity analyses.
}

\section{probabilistic}{

Methods for post-processing the results of probabilistic sensitivity analyses.
}

