% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{incremental}
\alias{incremental}
\title{Incremental cost-effectiveness analysis}
\usage{
incremental(cea, extended = T, subset = NULL)
}
\arguments{
\item{cea}{an object of class CEA for which the comparisons are being
caclulated}

\item{subset}{an optional vector specifying a subset of analyses to be
used.}
}
\value{
an object of class pairCEA
}
\description{
Performs incremental comparisons of cost-effectiveness given an existing
CEA object.
}

