% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perms.R
\name{independent}
\alias{independent}
\title{Independent model of strain acquisition}
\usage{
independent(params, data, pci_list, max_moi = 25)
}
\arguments{
\item{params}{Vector of parameters including the frequency of each species
and the mean moi mu.}

\item{data}{Real data to compare model predictions to.}

\item{pci_list}{Permutation, composition and infections list.}
}
\value{
Returns a named multinomial probability distribution.
}
\description{
Independent model of strain acquisition
}
\details{
For a given frequency of each strain and mean moi, returns the
  related multionomial distribution describing each infection type.
}
