% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_test.R
\name{likelihood_test}
\alias{likelihood_test}
\title{Model likelihood}
\usage{
likelihood_test(res, models, boot_iter = 50000, plot = TRUE,
  quantiles = c(0.025, 0.975), comp_func = dmultinom,
  scales = "fixed")
}
\arguments{
\item{res}{Result of \code{cooccurence_test}. Can be an individual model
or a named list of them.}

\item{boot_iter}{Bootstrap iterations. Default = 10000}

\item{plot}{Boolean for default plotting the bootsrap results. Default = TRUE}

\item{quantiles}{Vector of length 2 for the quantiles used. Default = `c(0.025, 0.975)`}
}
\value{
Invisibly returns a list containing the plot of our data
  compared to the boostrapped draws from the model.
}
\description{
Maximum likelihood test and plotted results for coinfection dynamics
}
\details{
Takes the result of the cooccurence test and assesses the
  overall fit of the model by drawing from the model and comparing
  to the likelihood of the data
}
\examples{
\dontrun{

# example of the two forms of data type accepted
real <- data.frame(
"variable"=c("pf/po","pf/pv","pf/po/pv","pf","po","po/pv","pv"),
"value"=c(84,179,1,5181,44,1,309)
)

real <- c(
"pf/po" = 84, "pf/pv" = 179, "pf/po/pv" = 1,
"pf" = 5181, "po" = 44, "po/pv" = 1, "pv" = 309
)

res <- cooccurence_test(data = real)
ll <- likelihood_test(res)

}

}
