% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{LIMMA_analysis}
\alias{LIMMA_analysis}
\title{Perform differential expression analysis using LIMMA}
\usage{
LIMMA_analysis(
  data,
  assignments = NULL,
  batch = NULL,
  tech_reps = NULL,
  contrast = NULL
)
}
\arguments{
\item{data}{Table of protein quantifications with samples in columns and features in rows.}

\item{assignments}{Character vector of annotations of grouping per sample. By default set to NULL. In this case an ordinary one-sample test is performed.}

\item{batch}{Optional character vector specifying sample batches.}

\item{tech_reps}{Optional character vector specifying if samples are technical replicates}

\item{contrast}{String of format Group1_vs_Group2 specifying contrast of interest. Replace Group1 and Group2 by groups specified in assignments.}
}
\value{
Returns a matrix which rows correspond to the proteins under analysis and columns indicate the corresponding abundance ratio, t-statistic, p-value and FDR adjusted p-value.
}
\description{
Perform differential expression analysis using LIMMA
}
\details{
Perform differential expression analysis using R-package LIMMA.
}
