% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{load_Requant_data}
\alias{load_Requant_data}
\title{Load IceR result files}
\usage{
load_Requant_data(
  path_to_requant_folder = NA,
  file_name_extension = NA,
  path_MaxQ = NA,
  quant_value = c("LFQ", "Total", "Top3"),
  min_feat_count = 1,
  min_feat_count_criteria = c("all", "unique"),
  imputed = T
)
}
\arguments{
\item{path_to_requant_folder}{Optional path to folder containing IceR outputs. By default set to NA. In this case a file browser is opened. If path is directly specified, it has to end with \\}

\item{file_name_extension}{Optional file name extension of IceR output files. Only required if path_to_requant_folder directly specified.}

\item{path_MaxQ}{Optional path to folder containing MaxQuant outputs. By default set to NA. In this case a file browser is opened. If path is directly specified, it has to end with \\}

\item{quant_value}{Specifying which protein quantification data should be used. Selection between "LFQ", "Total" or "Top3". By default set to "LFQ".}

\item{min_feat_count}{Minimal required number of quantified features per protein. By default set to 1.}

\item{min_feat_count_criteria}{Criteria how to count quantified features per protein. Either all or only unique peptide features are counted. By default set to "all".}

\item{imputed}{Boolean value indicating if data with noise model based imputation should be used. By default set to T.}
}
\value{
List object containing protein and peptide quantification information in sub-lists named Protein_level and Peptide_level, respectively.
}
\description{
Load IceR result files
}
\details{
Wrapper function to load and filter IceR results.
}
