% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{maxDensity}
\alias{maxDensity}
\title{Function to determine at which value a density maximum is reached}
\usage{
maxDensity(data)
}
\arguments{
\item{data}{Numeric vector}
}
\value{
Numeric indicating at which value a density maximum is reached
}
\description{
Function to determine at which value a density maximum is reached
}
\details{
Uses kernel density estimation function from R-package stats removing missing values
}
