% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{moi_probabilities}
\alias{moi_probabilities}
\title{Find the probabilities for the Mean MOI in the Population}
\usage{
moi_probabilities(obj, maxmoi, params)
}
\arguments{
\item{obj}{S4 object;}

\item{maxmoi}{numeric; maximum MOI considered}

\item{params}{parameter vector; Must contain at least one value named `mu`,
  which is the mean muliplicity of infection. Could also contain `size` in
  which case negative binomial distribution is used. See details.

 @details Using either a poisson or negative binomial distribution, the
  probability of drawing some series of infections is considered.
  In the case of where the true prevalence is known ("surveillance class),
  we consider the probabilities across the `0:nrow(multi_dens)` distribution.
  In the case that only the case distribution is known, we can only
  consider the probabilities across the `1:nrow(multi_dens)` distribution
  but will need to re-weight it to sum to 1.

  These are used as weights to then calculate the weighted mean
  probability of each infection type (the columns in `multi_dens`).}
}
\description{
Find the probabilities for the Mean MOI in the Population
}
