% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize quantification data}
\usage{
normalize_data(
  data,
  method = c("median", "density", "vsn"),
  norm_to = NULL,
  norm_on_subset = NULL,
  main = "Data"
)
}
\arguments{
\item{data}{Table of quantifications with samples in columns and features in rows.}

\item{method}{Method how data should be normalized. Select between "median","density","vsn". By default set to "median".}

\item{norm_to}{Optional: specify to which intensity all samples should be normalized. By default set to NULL indicating that samples are normalized to the average of samples.}

\item{norm_on_subset}{Optional: numeric vector of rows on which normalization factors should be determined (e.g. expected constant background)}

\item{main}{Titel of plots}
}
\value{
Table with normalized quantifications and density plots.
}
\description{
Normalize quantification data
}
\details{
Normalize quantification data
}
