% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{pairwise}
\alias{pairwise}
\title{Calculate pairwise comparisons}
\usage{
pairwise(cea, referent = NULL, subset = NULL)
}
\arguments{
\item{cea}{an object of class CEA for which the comparisons are being
caclulated}

\item{referent}{an optional value indicated which intervention is to be
used as the referent for pairwise comparisons.  If not specified, the first
intervention in each analysis will be designated the referent.}

\item{subset}{an optional vector specifying a subset of analyses to be
used.}
}
\value{
an object of class CEA.pairwise
}
\description{
Calculates pairwise comparisons for one or more analyses.
}

