% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{peak_decision}
\alias{peak_decision}
\title{Peak decision algorithm only designed for internal use}
\usage{
peak_decision(
  features_select,
  peak_quant,
  samples,
  s,
  RT_correction_factors,
  mz_correction_factors,
  features_intensity_sample,
  Ioncount_sample,
  feature_with_background_intensity_sample,
  Ioncount_with_background_sample,
  peak_selected_sample,
  delta_mz,
  delta_rt,
  peak_min_ion_count,
  chunk = NULL,
  num_chunks = NULL,
  progress = T
)
}
\arguments{
\item{features_select}{Features on which peak-selection should be performed}

\item{peak_quant}{List of determined peaks per feature}

\item{samples}{String vector of all samples}

\item{s}{Integer indicating index of current sample}

\item{RT_correction_factors}{RT corrections}

\item{mz_correction_factors}{mz corrections}

\item{features_intensity_sample}{features_intensity_sample}

\item{Ioncount_sample}{Ioncount_sample}

\item{feature_with_background_intensity_sample}{feature_with_background_intensity_sample}

\item{Ioncount_with_background_sample}{Ioncount_with_background_sample}

\item{peak_selected_sample}{peak_selected_sample}

\item{delta_mz}{delta_mz}

\item{delta_rt}{delta_rt}

\item{peak_min_ion_count}{peak_min_ion_count}

\item{chunk}{chunk}

\item{num_chunks}{num_chunks}

\item{progress}{Progressbar}
}
\description{
Peak decision algorithm only designed for internal use
}
\details{
Peak decision algorithm
}
