% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{Peak_selection_FDR}
\alias{Peak_selection_FDR}
\title{Peak selection FDR algorithm for internal use}
\usage{
Peak_selection_FDR(
  num_features = 500,
  features,
  samples,
  peaks,
  path_to_features,
  peak_quant,
  feature_with_background_intensity,
  peak_selected,
  delta_mz,
  delta_rt,
  peak_min_ion_count,
  num_peaks_store,
  alignment_scores_cutoff,
  n_cores,
  peak_decision,
  Alignment_scoring,
  seed = 110519,
  plot = T
)
}
\arguments{
\item{num_features}{num_features}

\item{features}{features}

\item{samples}{samples}

\item{peaks}{peaks}

\item{path_to_features}{path_to_features}

\item{peak_quant}{peak_quant}

\item{feature_with_background_intensity}{feature_with_background_intensity}

\item{peak_selected}{peak_selected}

\item{delta_mz}{delta_mz}

\item{delta_rt}{delta_rt}

\item{peak_min_ion_count}{peak_min_ion_count}

\item{num_peaks_store}{num_peaks_store}

\item{alignment_scores_cutoff}{alignment_scores_cutoff}

\item{n_cores}{n_cores}

\item{peak_decision}{peak_decision}

\item{Alignment_scoring}{Alignment_scoring}

\item{seed}{seed}

\item{plot}{plot}
}
\description{
Peak selection FDR algorithm for internal use
}
\details{
Peak selection FDR algorithm for internal use
}
