% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{PECA_analysis}
\alias{PECA_analysis}
\title{Perform differential expression analysis using peptide-level expression-change averaging}
\usage{
PECA_analysis(
  peptide_data,
  peptide_data_quant_significance = NULL,
  ids,
  anno = NULL,
  group1_name,
  group2_name,
  TopN_ratio = 5,
  pvalue_cutoff = NA,
  test = c("t", "modt", "rots")
)
}
\arguments{
\item{peptide_data}{Table of peptide quantifications with samples in columns and features in rows.}

\item{peptide_data_quant_significance}{Optional: Ion accumulation significances with samples in columns and features in rows., By default not required and set to NULL.}

\item{ids}{character vector of same length as rows in peptide_data indicating to which protein ID the corresponding peptide belongs to.}

\item{anno}{Annotation of grouping per sample}

\item{group1_name}{Name of group1 in annotation}

\item{group2_name}{Name of group1 in annotation}

\item{TopN_ratio}{Number of top abundant peptides on which at maximum protein ratios should be estimated. By default set to 5.}

\item{pvalue_cutoff}{Optional: Ion accumulation significance cutoff.}

\item{test}{A character string indicating whether the ordinary t-test ("t"), modified t-test ("modt"), or reproducibility-optimized test statistic ("rots") is performed.}
}
\value{
Returns a matrix which rows correspond to the genes under analysis and columns indicate the corresponding abundance ratio, t-statistic, p-value and FDR adjusted p-value
}
\description{
Perform differential expression analysis using peptide-level expression-change averaging
}
\details{
Perform differential expression analysis using the function PECA_df() from the R-package PECA.
}
