% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{Compare coefficients of variation of quantifications between data sets.}
\usage{
plot_accuracy(
  list_of_data_lists,
  colors = c("darkgrey", "chocolate3"),
  Legendpos = "topleft",
  margins = c(2, 4, 4, 10),
  inset = c(-0.435, 0),
  Annotation_column = 1,
  plot_for = c("both", "protein", "peptide"),
  allow_missing_values = F,
  representation_of = c("total", "per group", "per intensity"),
  show_numbers = T,
  numbers_size = 1,
  round_to_k = T
)
}
\arguments{
\item{list_of_data_lists}{List object containing lists of loaded MaxQuant or IceR data. Requires sample annotation to be added by function add_annotations().}

\item{colors}{Colors for data sets to be compared. By default two colors are specified ("darkgrey","chocolate3")}

\item{Legendpos}{Location of legend, By default set to "top"}

\item{margins}{Margins aroung plot area. By default set to c(12,4,4,9)}

\item{inset}{Inset of legend. By default set to c(-0.435,0)}

\item{Annotation_column}{Which annotation column should be used to determine between which samples variability should be determined. By default set to 1.}

\item{plot_for}{Specify if plots should be generated for protein-level ("protein"), peptide-level ("peptide") or both ("both"). By default set to "both".}

\item{allow_missing_values}{Indicate if missing values are allowed during variability estimation. By default set to F.}

\item{representation_of}{Indicate how results should be visualized. Select between "total","per group","per intensity". "total" = aggregate CVs into a single boxplot per data set. "per group" = aggregate CVs per annotation group per data set. "per intensity" = plot per data set CVs binned by protein/peptide abundance. By default set to "total".}

\item{show_numbers}{Indicate if counts of available CVs per boxplot should be plotted. By default set to T.}

\item{numbers_size}{Numeric value specifying plotting size of CV counts. By default set to 1.}

\item{round_to_k}{Indicate if CV counts should be rounded to 1000 (K). By default set to T.}
}
\value{
Comparison plots
}
\description{
Compare coefficients of variation of quantifications between data sets.
}
\details{
Compare CVs between data sets.
}
