% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{prepare_preprocessed_data}
\alias{prepare_preprocessed_data}
\title{Prepare preprocessed data for IceR}
\usage{
prepare_preprocessed_data(
  path_to_folder,
  path_to_output,
  pipeline = c("MaxQ", "User"),
  sample_list = NULL,
  MassSpec_mode = c("Orbitrap", "TIMSToF"),
  multiplicity = c(1, 2, 3),
  SILAC_settings = list(light = c(""), medium = c("Lys4", "Arg6"), heavy = c("Lys8",
    "Arg10")),
  remove_contaminants = T
)
}
\arguments{
\item{path_to_folder}{Path to folder containing the results from the preprocessing pipeline (e.g. MaxQuant txt folder containing at least allpeptides.txt, evidence.txt, peptides.txt and proteinGroups.txt)}

\item{path_to_output}{Path to folder where IceR results should be stored}

\item{pipeline}{Preprocessing pipeline. Can be one of the following: MaxQ, Unique. Unique indicates that data was rearranged already manually. Defaults to MaxQ.}

\item{sample_list}{Character vector (raw file names) listing which samples should be kept. By default all samples occurring in outputs are aligned.}

\item{MassSpec_mode}{String being either "Orbitrap" or "TIMSToF" specifying by which type of Mass Spectrometer the data was generated. By default it expects Thermo Orbitrap data.}

\item{multiplicity}{Numeric value between 1 - 3 indicating if multiple channels are mixed on MS1-level. Label-free corresponds to 1 and SILAC to 2 (light and heavy) or 3 (light, medium, heavy)}

\item{SILAC_settings}{List of 3 string vectors named light, medium and heavy. Each string vector should indicate which heavy isotope of Lys and/or Arg was used for respective channel. String vectors can contain Lys2, Lys4, Lys6, Lys8, Arg6 and Arg10.}

\item{remove_contaminants}{Boolean value indicating if peptide features labeled as contaminants should be removed. By default set to T.}
}
\value{
Rearranged data is stored in the defined output folder ready for next steps of IceR and is returned as list object
}
\description{
Prepare preprocessed data for IceR
}
\details{
Rearranges data from preprocess pipeline so that data can be used by IceR
}
