% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{prepare_rawTIMS}
\alias{prepare_rawTIMS}
\title{Prepare for access of raw TIMS-ToF Pro data}
\usage{
prepare_rawTIMS(path_to_raw = NULL)
}
\arguments{
\item{path_to_raw}{Path to folder containing d-files}
}
\description{
Prepare for access of raw TIMS-ToF Pro data
}
\details{
Getting prepared for acess of raw timsToF Pro data
}
