% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{requantify_features}
\alias{requantify_features}
\title{Perform quantification of IceR features}
\usage{
requantify_features(
  path_to_features,
  path_to_mzXML = NA,
  path_to_input,
  feature_table_file_name = "Features_aligned_merged_IceR_analysis.txt",
  output_file_names_add = "IceR_analysis",
  RT_calibration = T,
  mz_calibration = T,
  abundance_estimation_correction = T,
  Quant_pVal_cut = 0.05,
  n_cores = 2,
  kde_resolution = 50,
  num_peaks_store = 5,
  plot_peak_detection = F,
  alignment_variability_score_cutoff = 0.05,
  alignment_scores_cutoff = 0.05,
  mono_iso_alignment_cutoff = 0.05,
  calc_peptide_LFQ = F,
  calc_protein_LFQ = T,
  MassSpec_mode = c("Orbitrap", "TIMSToF"),
  use_IM_data = T,
  path_to_extracted_spectra = NA
)
}
\arguments{
\item{path_to_features}{Path to folder where results of align_features() are stored}

\item{path_to_mzXML}{Path to folder containing mzXML files of samples in case of Orbitrap data.}

\item{path_to_input}{Path to folder containing MaxQuant outputs (txt folder containing at least allpeptides.txt, evidence.txt, peptides.txt and proteinGroups.txt)}

\item{feature_table_file_name}{File name which contains align_features() results. By default is set to Features_aligned_merged_IceR_analysis.txt.}

\item{output_file_names_add}{IceR result name tag. By default IceR_analysis}

\item{RT_calibration}{Boolean value indicating if corrected RT should be used during peak detection, selection and DICE, By default set to T.}

\item{mz_calibration}{Boolean value indicating if corrected m/z should be used during peak detection, selection and DICE, By default set to T.}

\item{abundance_estimation_correction}{Boolean value indicating if resulting peptide abundances should be corrected using MaxQuant results as a reference. By default set to T.}

\item{Quant_pVal_cut}{Numeric value used as diagnostic cutoff border for visualization of significances of ion accumulation per IceR feature quantification. Furthermore, used as cutoff to filter +1-isotopic IceR features with significant accumulation of ions. By default set to 0.05.}

\item{n_cores}{Numeric value specifying on how many CPU cores tasks should be parallelized. By default set to 2.}

\item{kde_resolution}{Numeric value specifying number of grid points per dimension. By default set to 50.}

\item{num_peaks_store}{Numeric value specifying number of 2D peaks to be stored during peak detection. By default set to 5.}

\item{plot_peak_detection}{Boolean value indicating if for every feature quantification the determined kernel density estimations and detected peaks should be visualized and stored. By default set to F.}

\item{alignment_variability_score_cutoff}{Numeric value specifying significance cutoff to distinguish which features show high RT- or m/z-variability of selected peaks between samples. By default set to 0.05. All features showing significant general variability (variability score < alignment_variability_score_cutoff) are excluded.}

\item{alignment_scores_cutoff}{Numeric value specifying significance cutoff to distinguish which samples show high RT- or m/z-variability of selected peaks for respective IceR feature. By default set to 0.05. All samples showing significant peak variability for respective IceR feature (variability score < alignment_scores_cutoff) are excluded (quantification set to NA).}

\item{mono_iso_alignment_cutoff}{Numeric value specifying significance cutoff to distinguish which samples show high RT- or m/z-variability of selected peaks for +1-isotopic from corresponding monoisotopic IceR feature. By default set to 0.05. All samples showing significant peak variability between selected +1-isotopic and monoisotopic IceR features (variability score < mono_iso_alignment_cutoff) are excluded (quantification of +1-isotopic feature set to NA).}

\item{calc_peptide_LFQ}{Boolean value specifying if multiply peptide quantification data for same peptide sequence (multiply charge states, isotope-states) should be aggregated using the MaxLFQ algorithm. By default set to F.}

\item{calc_protein_LFQ}{Boolean value specifying if protein quantification should be additionally performed by peptide quantification aggregation using the MaxLFQ algorithm. By default set to T.}

\item{MassSpec_mode}{String being either "Orbitrap" or "TIMSToF" specifying by which type of Mass Spectrometer the data was generated. By default it expects Thermo Orbitrap data.}

\item{use_IM_data}{Boolean value indicating if ion mobility information should be used during feature quantification in case of TIMS-ToF data. By default set to T.}

\item{path_to_extracted_spectra}{Path to folder containing extracted spectra files of samples in case of TIMS-ToF data.}
}
\value{
Outputs are stored in the specified output folder and intermediate results are stored in the sub-directory Temporary_files.
Quantification results are stored in tab-delimited text files (.tab):
Feature information - Features_DDAiceR_Analysis.tab
Peak alignment scores - Features_quantification_alignment_score_DDAiceR_Analysis.tab
Feature quantifications - Features_quantification_DDAiceR_Analysis.tab
Feature quantifications after imputation - Features_quantification_imputed_DDAiceR_Analysis.tab
Numbers of observed ions per feature quantification - Features_quantification_ioncount_DDAiceR_Analysis.tab
Alignment scores between monoisotopic and corresponding +1isotopic IceR features - Features_quantification_mono_iso_alignment_score_DDAiceR_Analysis.tab
Significance of ion accumulations per feature quantification - Features_quantification_pvals_DDAiceR_Analysis.tab
Signal to background intensity ratios per feature quantification - Features_quantification_S2B_DDAiceR_Analysis.tab
General variability score of peak selections - Features_quantification_variability_score_DDAiceR_Analysis.tab
Protein quantification - Proteins_quantification_LFQ_DDAiceR_Analysis.tab
Protein quantification after imputation - Proteins_quantification_LFQ_imputed_DDAiceR_Analysis.tab
QC results of background noise estimations are visualized in "Decoy feature quantification parameters.pdf".
QC results of peak selections are visualized in "Alignment and quantification scores.pdf".
The performance of RT- and m/z-alignments over samples is visualized in "Performance of feature alignment.pdf"
Estimation of required peptide abundance correction factors is visualized in "Correct feature abundance estimations Signal_Background_intensity.pdf".
Intermediate results of the function are stored in RData files.
}
\description{
Perform quantification of IceR features
}
\details{
Performs final steps of the IceR workflow:
1) Estimation of background noise per IceR feature quantification.
2) 2D Kernel density estimation-based peak detection, selection and DICE-based quantification of IceR features.
3) Determination of significances of ion accumulations per IceR feature quantification.
4) Quality control of peak selections.
5) IceR peak selection accuracy estimations per sample.
6) Optional: Imputation of missing IceR feature quantifications using estimated backgroudn noise models per sample.
7) Protein quantification by aggregating available peptide quantifications.
}
