% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature_alignment_requantification.R
\name{run_msconvert_raw_mzXML}
\alias{run_msconvert_raw_mzXML}
\title{Convert thermo raw files to mzXML with centroided ms1 scans using the ProteoWizard tool msConvert}
\usage{
run_msconvert_raw_mzXML(path_to_raw = NULL)
}
\arguments{
\item{path_to_raw}{Path to folder containing raw files which should be converted}
}
\value{
Resulting mzXML files are stored in a sub-directory within specified raw file folder
}
\description{
Convert thermo raw files to mzXML with centroided ms1 scans using the ProteoWizard tool msConvert
}
\details{
Requires installation of ProteoWizard (http://proteowizard.sourceforge.net/download.html). Pay attention to installation requirements.
}
