% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{SaveExcel}
\alias{SaveExcel}
\title{Save data.frames or list of data.frames as xlsx-files.}
\usage{
SaveExcel(Data, File)
}
\arguments{
\item{Data}{Object of type data.frame or list}

\item{File}{File name of xlsx-file}
}
\value{
xlsx-file
}
\description{
Save data.frames or list of data.frames as xlsx-files.
}
