% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{set_sample_names}
\alias{set_sample_names}
\title{Change sample (column) names of loaded MaxQuant or IceR data}
\usage{
set_sample_names(data_list, sample_names)
}
\arguments{
\item{data_list}{List object containing loaded MaxQuant or IceR data}

\item{sample_names}{Character vector of new names of same length as number of samples in MaxQuant or IceR data.}
}
\value{
List object containing loaded MaxQuant or IceR data with updated sample names.
}
\description{
Change sample (column) names of loaded MaxQuant or IceR data
}
\details{
Change sample names.
}
