% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{surveillance-class}
\alias{surveillance-class}
\title{Survelliance Class}
\description{
Survelliance Class
}
\details{
Class for surveillance data, or when the the true prevalence
is known. For a true prevalence to be known, we must have the
population at risk as the denominator (e.g. all individuals
that could have been sampled).
}
