% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tabulate}
\alias{tabulate}
\title{tabulates a solution response (requires the base request)}
\usage{
tabulate(solResponse, solveRequest)
}
\arguments{
\item{solResponse}{The solution response from the api.}

\item{solveRequest}{The base-model which was solved (i.e. submitted to the api).}
}
\description{
tabulates a solution response (requires the base request)
}
