% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perms.R
\name{cooccurence_test}
\alias{cooccurence_test}
\title{Co-occurence test}
\usage{
cooccurence_test(data, max_moi = 25, boot_iter = 10000)
}
\arguments{
\item{data}{Observed data. Either as a named vector or a 2 column
\code{\link{data.frame}}. See examples for more info.}

\item{max_moi}{Maxuimum infction composition explored. Default = 25.}

\item{boot_iter}{Bootstrap iterations. Default = 10000}
}
\value{
Invisibly returns a list containing the estimated parameters
  and a plot of our data compared to the boostrapped estimates.
}
\description{
Maximum likelihood test and plotted results for coinfection dynamics
}
\details{
Estimates the maximum likely population frequency of each species
  (i.e. the names of the observed entities in our data) and the mean number
  of infections given our observed data. These estimates are used to estimate
  the multinomial probability distribution for all cooccurences and comparing
  these to our data using a boostrap method.
}
\examples{
\dontrun{

# example of the two forms of data type accepted
real <- data.frame(
"variable"=c("pf/po","pf/pv","pf/po/pv","pf","po","po/pv","pv"),
"value"=c(84,179,1,5181,44,1,309)
)

real <- c(
"pf/po" = 84, "pf/pv" = 179, "pf/po/pv" = 1,
"pf" = 5181, "po" = 44, "po/pv" = 1, "pv" = 309
)

res <- cooccurence_test(data = real)

}
}
