% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perms.R
\name{generate_composition_levels}
\alias{generate_composition_levels}
\title{Determines the potential compositions from the cases names
(e.g. composition levels)}
\usage{
generate_composition_levels(names)
}
\arguments{
\item{names}{character vector of case names}
}
\value{
Returns a vector of composition levels
}
\description{
Determines the potential compositions from the cases names
(e.g. composition levels)
}
\details{
Creates a named vector of all combinations for the names
  in probs, separated by a "/" and sorted by character length
  and then alphabetically
}
\examples{
\dontrun{

obj <- new("casedetect")
obj@cases <- c(50, 50, 100)
obj@casenames <- c("a", "b", "c")

icer:::generate_composition_levels(names = obj@casenames)

#> a     b     c   a/b   a/c   b/c a/b/c

}
}
