% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{icer}
\alias{icer}
\title{Calculate incremental cost-effectiveness ratio}
\usage{
icer(deltaEff, deltaCost)
}
\arguments{
\item{deltaEff}{Differences in clinical effectiveness.}

\item{deltaCost}{Differences in costs.}
}
\value{
An object of class ICER.
}
\description{
Calculate the incremental cost-effectiveness ratio of one or more interventions based on the
differences in clinical effectiveness and costs.
}

