% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{icer}
\alias{icer}
\title{Calculate incremental cost-effectiveness ratios for a given set of analyses}
\usage{
icer(cea)
}
\arguments{
\item{pairCEA}{An object of class pairCEA}
}
\value{
An object of class pairICER.
}
\description{
Calculate the incremental cost-effectiveness ratio of one or more interventions based on the
differences in clinical effectiveness and costs.  Depending on
}

