% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{icer}
\alias{icer}
\title{Caclulate incremental cost-effectiveness ratio based on differenes in effectiveness, cost}
\usage{
icer(deltaEff, deltaCost)
}
\arguments{
\item{deltaEff}{Differences in clinical efficacy.}

\item{deltaCost}{Differences in cost.}
}
\value{
An object of class ICER.
}
\description{
Caclulate incremental cost-effectiveness ratio based on differenes in effectiveness, cost
}
\details{
ICER objects are numeric vectors representing cost-effectiveness ratios.  When
the referent is more costly/more effective than the comparator, the ICER is represented
as a postive number.  When the comparator is more costly/more effective, the ICER is
represented as a negative value.  Dominance is represented by infinite values (\code{-Inf}: Dominant,
\code{Inf}: Dominated).  If the referent have exactly as costly/effective, the ICER is represented as
\code{NaN}.
}

