% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{icer}
\alias{icer}
\alias{icer-package}
\title{\pkg{icer} Infection compistion estimation}
\description{
Maximum likelihood estimator for the occurence
  of mixed infections given data on mixed infections.
}
